backend/requirements.txt
fastapi>=0.111,<1.0
pydantic>=2,<3
sqlalchemy>=2,<3
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0

